Page tag now honors "without-aside". Asides with "0" size/width are not rendered.
The Page DRYML tag has been updated so that when "without-aside" is
used as an attribute in a <page> tag, the aside is actually removed
from the markup rather than just being rendered with no contents (a
blank column). Also, if the aside-size attribute is set to 0, the aside
is also removed from the markup and the main content is given all
available width (span12). Also updated the comments to add more
detail/clarity.